requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.4
boto3==1.34.159
botocore==1.34.159
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpickle==1.6.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.7
# dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
# distro==1.7.0
# distro-info==1.1+ubuntu0.2
docker==7.1.0
durationpy==0.9
ecdsa==0.19.0
editor==1.6.6
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.20.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
inquirer==3.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
# jupyter-events==0.10.0
# jupyter-lsp==2.2.5
# jupyter-server-mathjax==0.2.6
# jupyter_client==8.6.2
# jupyter_core==5.7.2
# jupyter_server==2.14.2
# jupyter_server_terminals==0.5.3
# jupyterlab==4.2.4
# jupyterlab_git==0.50.1
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.27.3
# jupyterlab_widgets==3.0.11
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langchain-upstage==0.6.0
langgraph==0.3.22
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.22
# launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.3.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbdime==4.0.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook_shim==0.2.4
numpy>=1.26.2,<2.0.0
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.20.5
# nvidia-nvjitlink-cu12==12.1.105
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.70.0
opencv-python==4.10.0.84
# opentelemetry-api==1.31.1
# opentelemetry-exporter-otlp-proto-common==1.31.1
# opentelemetry-exporter-otlp-proto-grpc==1.31.1
# opentelemetry-instrumentation==0.52b1
# opentelemetry-instrumentation-asgi==0.52b1
# opentelemetry-instrumentation-fastapi==0.52b1
# opentelemetry-proto==1.31.1
# opentelemetry-sdk==1.31.1
# opentelemetry-semantic-conventions==0.52b1
# opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging<24,>=16.8
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
posthog==3.23.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
proto-plus==1.24.0
protobuf<5,>=3.20
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.8.1
pydantic_core==2.20.1
Pygments==2.18.0
# PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pynvml==11.5.3
pyparsing==2.4.7
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
# python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
readchar==4.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich>=10.14.0,<14.0.0
rpds-py==0.20.0
rsa==4.9
runs==1.2.2
s3transfer==0.10.2
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
sentry-sdk==2.12.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spinners==0.0.24
SQLAlchemy==2.0.40
ssh-import-id==5.11
sshpubkeys==3.3.1
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity>=8.2.3,<9.0.0
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tiktoken==0.9.0
timeago==1.0.16
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
# torch==2.3.1+cu121
# torchaudio==2.3.1+cu121
# torchvision==0.18.1+cu121

torch
torchaudio
torchvision
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
# triton==2.3.1
typer==0.15.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
# unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.21.0
vessl==0.1.192
# wadllib==1.3.6
watchdog==3.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.11
wrapt==1.17.2
xmod==1.8.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
streamlit==1.32.0
faiss-cpu==1.9.0
python-dotenv==1.0.1
selenium
webdriver-manager